requirements.txt
selenium
Pillow
requests